backend/requirements.txt
# cSpell:ignore djangorestframework chromadb
Django==5.1.6
djangorestframework==3.15.2
stripe==12.5.1
django-cors-headers==4.6.0
python-decouple==3.8
chromadb==0.4.22
sentence-transformers==2.3.1
PyPDF2==3.0.1
requests==2.31.0
torch>=2.0.0